+++ Editor/MonoFluxEditor.cs
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2023 Xavier Arpa López Thomas Peter ('Kingdox')
+Copyright (c) 2023 Xavier Arpa López Thomas Peter ('xavierarpa')
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -25,7 +25,7 @@ using System.Reflection;
 using System;
 using System.Linq;
 using System.Collections.Generic;
-namespace Kingdox.UniFlux.Editor
+namespace UniFlux.Editor
 {
     [CustomEditor(typeof(MonoFlux), true)]
     public partial class MonoFluxEditor : UnityEditor.Editor